+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 All notable changes to this project are documented in this file.
 
+## v1.0 - *TBD*
+
+### payouts-substitutor
+
+- Refactored the logging process
+
 ## v0.2 - 2024-06-19
 
 ### General